+++ CHANGELOG.md
@@ -16,6 +16,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Added iconic style for large Button.
 
+- Added Popover component.
+
 ## [0.27.0] - 2020-10-14
 
 - Couple of small updates in Tooltip and Modal.